requirements.txt
apify>=2.0.0
google-generativeai>=0.8.0
openai>=1.50.0
httpx>=0.27.0
playwright>=1.40.0
Pillow>=10.0.0
